Breakpoint split view mouse hovers
