+++ plugins/breakpoint-split-view/src/BreakpointSplitView/components/Overlay.tsx
@@ -15,14 +15,15 @@ const Overlay = observer(function (props: {
 }) {
   const { model, trackId } = props
   const tracks = model.getMatchedTracks(trackId)
+  const type = tracks[0]?.type
 
   // curvy line type arcs
-  if (tracks[0]?.type === 'AlignmentsTrack') {
+  if (type === 'AlignmentsTrack') {
     return <AlignmentConnections {...props} />
   }
 
   // translocation type arcs
-  else if (tracks[0]?.type === 'VariantTrack') {
+  else if (type === 'VariantTrack') {
     return model.hasTranslocations(trackId) ? (
       <Translocations {...props} />
     ) : model.hasPairedFeatures(trackId) ? (
@@ -30,7 +31,10 @@ const Overlay = observer(function (props: {
     ) : (
       <Breakends {...props} />
     )
-  } else {
+  }
+
+  // unknown
+  else {
     return null
   }
 })
