+++ plugins/breakpoint-split-view/src/BreakpointSplitView/components/PairedFeatures.tsx
@@ -22,7 +22,7 @@ const PairedFeatures = observer(function ({
   parentRef: React.RefObject<SVGSVGElement | null>
   getTrackYPosOverride?: (trackId: string, level: number) => number
 }) {
-  const { views } = model
+  const { interactiveOverlay, views } = model
   const session = getSession(model)
   const { assemblyManager } = session
   const totalFeatures = model.getTrackFeatures(trackId)
@@ -99,6 +99,7 @@ const PairedFeatures = observer(function ({
               d={path}
               data-testid="r2"
               key={JSON.stringify(path)}
+              pointerEvents={interactiveOverlay ? 'auto' : undefined}
               strokeWidth={id === mouseoverElt ? 10 : 5}
               onClick={() => {
                 const featureWidget = session.addWidget?.(
