+++ plugins/breakpoint-split-view/src/BreakpointSplitView/types.ts
@@ -1,0 +1,22 @@
+export interface ExportSvgOptions {
+  rasterizeLayers?: boolean
+  filename?: string
+  Wrapper?: React.FC<{ children: React.ReactNode }>
+  fontSize?: number
+  rulerHeight?: number
+  textHeight?: number
+  paddingHeight?: number
+  headerHeight?: number
+  cytobandHeight?: number
+  trackLabels?: string
+  themeName?: string
+}
+
+export interface Breakend {
+  MateDirection: string
+  Join: string
+  Replacement: string
+  MatePosition: string
+}
+
+export type LayoutRecord = [number, number, number, number]
