+++ plugins/breakpoint-split-view/src/BreakpointSplitView/util.ts
@@ -1,8 +1,11 @@
 import { useEffect, useState } from 'react'
 
-import { clamp } from '@jbrowse/core/util'
+import { getConf } from '@jbrowse/core/configuration'
+import { clamp, getSession } from '@jbrowse/core/util'
 
-import type { LayoutRecord } from './model'
+import type { LayoutRecord } from './types'
+import type { AnyConfigurationModel } from '@jbrowse/core/configuration'
+import type { Feature } from '@jbrowse/core/util'
 import type { LinearGenomeViewModel } from '@jbrowse/plugin-linear-genome-view'
 
 type LGV = LinearGenomeViewModel
@@ -11,6 +14,7 @@ interface Display {
   height: number
   scrollTop: number
   SNPCoverageDisplay?: { height: number }
+  searchFeatureByID?: (str: string) => LayoutRecord
 }
 
 interface Track {
@@ -86,4 +90,45 @@ export function intersect<T>(
   const ids = new Set(a2.map(elt => cb(elt)))
   const a12 = a1.filter(value => ids.has(cb(value)))
   return rest.length === 0 ? a12 : intersect(cb, a12, ...rest)
+}
+
+const startClip = new RegExp(/(\d+)[SH]$/)
+const endClip = new RegExp(/^(\d+)([SH])/)
+
+export function getClip(cigar: string, strand: number) {
+  return strand === -1
+    ? +(startClip.exec(cigar) || [])[1]! || 0
+    : +(endClip.exec(cigar) || [])[1]! || 0
+}
+
+export function calc(track: Track, f: Feature) {
+  return track.displays[0]!.searchFeatureByID?.(f.id())
+}
+
+export async function getBlockFeatures(
+  model: { views: LinearGenomeViewModel[] },
+  track: { configuration: AnyConfigurationModel },
+) {
+  const { views } = model
+  const { rpcManager, assemblyManager } = getSession(model)
+  const assemblyName = model.views[0]?.assemblyNames[0]
+  if (!assemblyName) {
+    return undefined
+  }
+  const assembly = await assemblyManager.waitForAssembly(assemblyName)
+  if (!assembly) {
+    return undefined // throw new Error(`assembly not found: "${assemblyName}"`)
+  }
+  const sessionId = track.configuration.trackId
+  return Promise.all(
+    views.map(async view =>
+      (
+        (await rpcManager.call(sessionId, 'CoreGetFeatures', {
+          adapterConfig: getConf(track, ['adapter']),
+          sessionId,
+          regions: view.staticBlocks.contentBlocks,
+        })) as Feature[][]
+      ).flat(),
+    ),
+  )
 }
